Snake_Game by using Python
The Snake game is a nostalgic and addictive game where players control a snake that moves around the screen, consuming food to grow longer. The objective is to guide the snake towards the food while avoiding collisions with the walls or the snake's own body. As the snake consumes food, it increases in length, making the game progressively challenging.